requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.2
cffi==1.15.0
colorama==0.4.4
decorator==5.1.1
Django==2.2
django-ckeditor==6.4.2
django-js-asset==2.0.0
executing==0.8.3
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.2.2.post1
tzdata==2022.1
wcwidth==0.2.5
